Move management clients into management package
This commit moves around the packages for the Azure SDK as discussed in
'management', and storage has been moved to a top level concept.
