+++ .travis.yml
@@ -1,3 +1,3 @@
 language: go
 go: tip
-script: go test -v ./azure/...
+script: go test -v ./storage/...
