+++ management/virtualmachinedisk/entities.go
@@ -1,8 +1,8 @@
 package virtualmachinedisk
 
-import "github.com/MSOpenTech/azure-sdk-for-go/azure"
+import "github.com/MSOpenTech/azure-sdk-for-go/management"
 
 //DiskClient is used to manage operations on Azure Disks
 type DiskClient struct {
-	client azure.Client
+	client management.Client
 }
