+++ storage/client.go
